Title: 특정 문자로 시작하는 문자열, Time: 106ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240307/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -22,7 +22,7 @@ public class Main {
             }
         }
 
-        double strLengAvg = strLeng / cnt;
+        double strLengAvg = (double) strLeng / cnt;
         System.out.printf("%d %.2f", cnt, strLengAvg);
 
         br.close();
